+++ rng_rava/__init__.py
@@ -4,6 +4,8 @@ Copyright (c) 2023 Gabriel Guerrer
 Distributed under the MIT license - See LICENSE for details
 """
 
+__version__ = '1.2.0'
+
 from .rava_defs import *
 from .rava_rng import *
 from .rava_rng_aio import RAVA_RNG_AIO